+++ src/main/groovy/com/github/choonchernlim/choonchernlimArchetypeWebapp/config/SecurityConfig.groovy
@@ -1,7 +1,7 @@
 package com.github.choonchernlim.choonchernlimArchetypeWebapp.config
 
 import org.springframework.context.annotation.Configuration
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder
+import org.springframework.http.HttpMethod
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.builders.WebSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
@@ -11,26 +11,18 @@ import org.springframework.security.config.annotation.web.configuration.WebSecur
 @EnableWebSecurity
 class SecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
-    protected void configure(final AuthenticationManagerBuilder auth) throws Exception {
-    }
-
-    @Override
     protected void configure(final HttpSecurity http) throws Exception {
-        http.headers().
-                frameOptions().
-                sameOrigin(). // handling clickjacking prevention by setting X-FRAME-OPTIONS to SAMEORIGIN.
+        http.
+                headers().
+                frameOptions().deny(). // set X-FRAME-OPTIONS to DENY to prevent clickjacking attacks
                 and().
                 authorizeRequests().
-                antMatchers('/**').
-                permitAll()
+                antMatchers(HttpMethod.OPTIONS, '/**').denyAll(). // disable OPTIONS method to prevent XST attacks
+                antMatchers('/**').permitAll()
     }
 
     @Override
     void configure(final WebSecurity web) throws Exception {
-        web.ignoring().
-                antMatchers('/css/**').
-                antMatchers('/font/**').
-                antMatchers('/img/**').
-                antMatchers('/js/**')
+        web.ignoring().antMatchers('/assets/**')
     }
 }